Added action 'Add row to CSV'
