--- Tests.Utilities/Input/test.csv
@@ -1,12 +1,0 @@
-1,One (1),"One,","One (1)
-1","One,
-1"
-2,Two (2),"Two,","Two (2)
-2","Two,
-2"
-3,Three (3),"Three,","Three (3)
-3","Three,
-3"
-4,Four (4),"Four,","Four (4)
-4","Four,
-4"
