+++ Utilities/Apps.Utilities.csproj
@@ -5,7 +5,7 @@
     <Nullable>enable</Nullable>
     <Product>Utilities</Product>
     <Description>Various utility actions to parse or convert data</Description>
-    <Version>1.5.33</Version>
+    <Version>1.5.34</Version>
     <AssemblyName>Apps.Utilities</AssemblyName>
   </PropertyGroup>
   <ItemGroup>
